refactor: StockResponse 가격관련 필드 Int -> Long으로 수정

+++ src/main/kotlin/team/mozu/dsm/adapter/in/team/dto/response/StockResponse.kt
@@ -6,11 +6,11 @@ data class StockResponse(
     val id: UUID,
     val itemId: Int,
     val itemName: String,
-    val avgPurchasePrice: Int,
+    val avgPurchasePrice: Long,
     val quantity: Int,
-    val totalMoney: Int,
-    val nowMoney: Int,
-    val valuationMoney: Int,
+    val totalMoney: Long,
+    val nowMoney: Long,
+    val valuationMoney: Long,
     val valProfit: Int,
     val profitNum: Double
 )
